pyproject.toml
[tool.poetry]
name = "mead-website"
version = "0.1.0"
description = ""
authors = ["Alexander Mead <alexander.j.mead@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "^1.24.3"
scipy = "^1.10.1"
matplotlib = "^3.7.1"
camb = "^1.4.0"
hmcode = "^1.1.0"
numba = "^0.57.0"
hydra-core = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
